Added target for building tests
The 'test' target of the Makefile generated by CMake does not build the required
tests. Thus, a 'build-test' target has been added which does the job.
Removing the test build from the 'all' target should speed up the build of the
entire package dramatically which is in particular important for packaging.