Made single configuration builds mandatory for all platforms
This is a safe way how to get the build done as the user must decide
what he wants at configuration time.
Maybe the build system should provide some batch files to execute the build
and the test with the correct configuration set to avoid mistakes on the
command line.